move media-browser to shared module, fix some style mistakes
