+++ src/app/features/editor/editor.component.scss
@@ -14,6 +14,7 @@
 .container {
   display: flex;
   height: 100%;
+  max-width: 95%;
   padding: 1rem;
 }
