+++ src/app/shared/components/media-browser/media-browser.component.html
